% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{swap}
\alias{swap}
\title{Assign a Value to an Existing Name inside Another Package}
\usage{
swap(pkg, name, value)
}
\arguments{
\item{pkg}{name of package; \if{html}{\out{<character>}}}

\item{name}{name of the object; \if{html}{\out{<character>}}}

\item{value}{value to be bound to the object; \if{html}{\out{<any>}}}
}
\value{
invisible(value); \if{html}{\out{<any>}}
}
\description{
... replacing what is already there.
You must restart R in order reload the original.
}
\examples{
if (interactive()) {
snooze <- function() cat("zzz")
cuckoo::swap("utils", "alarm", snooze)
}
}
